Remove "impl" package in leshan-server-cf
