+++ leshan-bsserver-demo/src/main/java/org/eclipse/leshan/server/bootstrap/demo/servlet/ServerServlet.java
@@ -29,7 +29,7 @@ import org.apache.commons.lang.StringUtils;
 import org.eclipse.leshan.server.bootstrap.demo.json.PublicKeySerDes;
 import org.eclipse.leshan.server.bootstrap.demo.json.SecuritySerializer;
 import org.eclipse.leshan.server.bootstrap.demo.json.X509CertificateSerDes;
-import org.eclipse.leshan.server.californium.impl.LeshanBootstrapServer;
+import org.eclipse.leshan.server.californium.bootstrap.LeshanBootstrapServer;
 import org.eclipse.leshan.server.security.SecurityInfo;
 
 import com.eclipsesource.json.JsonObject;
