+++ leshan-server-cf/src/main/java/org/eclipse/leshan/server/californium/RootResource.java
@@ -13,7 +13,7 @@
  * Contributors:
  *     Sierra Wireless - initial API and implementation
  *******************************************************************************/
-package org.eclipse.leshan.server.californium.impl;
+package org.eclipse.leshan.server.californium;
 
 import java.util.List;
