+++ leshan-server-cf/src/main/java/org/eclipse/leshan/server/californium/observation/ObserveUtil.java
@@ -13,7 +13,7 @@
  * Contributors:
  *     Sierra Wireless - initial API and implementation
  *******************************************************************************/
-package org.eclipse.leshan.server.californium;
+package org.eclipse.leshan.server.californium.observation;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -24,6 +24,7 @@ import org.eclipse.leshan.core.node.LwM2mPath;
 import org.eclipse.leshan.core.observation.Observation;
 import org.eclipse.leshan.core.request.ContentFormat;
 import org.eclipse.leshan.core.request.ObserveRequest;
+import org.eclipse.leshan.server.californium.registration.CaliforniumRegistrationStore;
 
 /**
  * Utility functions to help to handle observation in Leshan. Those helper functions are only needed if you're
